Move macOS test to its own file
